Add tests to restart the docker and verify volume refcounts are restored.
Updated to modify some new functions

+++ tests/e2e/restart_test.go
@@ -21,39 +21,59 @@ package e2e
 
 import (
 	. "gopkg.in/check.v1"
+	"strings"
+	"log"
 
 	"github.com/vmware/docker-volume-vsphere/tests/utils/dockercli"
 	"github.com/vmware/docker-volume-vsphere/tests/utils/inputparams"
 	"github.com/vmware/docker-volume-vsphere/tests/utils/verification"
 	"github.com/vmware/docker-volume-vsphere/tests/utils/misc"
+	adminconst "github.com/vmware/docker-volume-vsphere/tests/constants/admincli"
+	adminutil "github.com/vmware/docker-volume-vsphere/tests/utils/admincli"
 )
 
 type RestartTestData struct {
 	config        *inputparams.TestConfig
 	volumeName    string
-	containerName string
+	containerName []string
 }
 
 func (s *RestartTestData) SetUpSuite(c *C) {
 	s.config = inputparams.GetTestConfig()
 	if s.config == nil {
+		log.Printf("Setup skipped")
 		c.Skip("Unable to retrieve test config, skipping docker/plugin restart tests")
 	}
 
 	s.volumeName = inputparams.GetUniqueVolumeName("restart_test")
-	s.containerName = inputparams.GetUniqueContainerName("restart_test")
+	s.containerName = []string{inputparams.GetUniqueContainerName("restart_test"),
+				inputparams.GetUniqueContainerName("restart_test"),
+				inputparams.GetUniqueContainerName("restart_test")}
 
 	// ensure there are no remanants from an earlier run
-	out, err := dockercli.RemoveContainer(s.config.DockerHosts[1], s.containerName)
+	out, err := dockercli.RemoveContainer(s.config.DockerHosts[1], s.containerName[0])
 
 	// Create a volume
 	out, err = dockercli.CreateVolume(s.config.DockerHosts[1], s.volumeName)
 	c.Assert(err, IsNil, Commentf(out))
+
+	// Add access for the second datastore to the VM's vmgroup
+	out, err = adminutil.AddDatastoreToVmgroup(s.config.EsxHost, adminconst.DefaultVMgroup, s.config.Datastores[1])
+	c.Assert(err, IsNil, Commentf(out))
+
+	log.Printf("Restart tests setup complete")
 }
 
 func (s *RestartTestData) TearDownSuite(c *C) {
-	out, err := dockercli.DeleteVolume(s.config.DockerHosts[1], s.volumeName)
-	c.Assert(err, IsNil, Commentf(out))
+	if s.config != nil {
+		out, err := dockercli.DeleteVolume(s.config.DockerHosts[1], s.volumeName)
+		c.Assert(err, IsNil, Commentf(out))
+
+		// Remove access for second datastore from the default vmgroup
+		out, err = adminutil.RemoveDatastoreFromVmgroup(s.config.EsxHost, adminconst.DefaultVMgroup, s.config.Datastores[1])
+		c.Assert(err, IsNil, Commentf(out))
+	}
+	log.Printf("Restart tests teardown complete")
 }
 
 var _ = Suite(&RestartTestData{})
@@ -72,7 +92,7 @@ func (s *RestartTestData) TestVolumeDetached(c *C) {
 	misc.LogTestStart(c.TestName())
 
 	// 1. Attach a volume to a container on a host
-	out, err := dockercli.AttachVolume(s.config.DockerHosts[1], s.volumeName, s.containerName)
+	out, err := dockercli.AttachVolume(s.config.DockerHosts[1], s.volumeName, s.containerName[0])
 	c.Assert(err, IsNil, Commentf(out))
 
 	// 2. Verify attached status
@@ -87,11 +107,11 @@ func (s *RestartTestData) TestVolumeDetached(c *C) {
 	status = verification.VerifyDetachedStatus(s.volumeName, s.config.DockerHosts[1], s.config.EsxHost)
 	c.Assert(status, Equals, true, Commentf("Volume %s is still attached", s.volumeName))
 
-	out, err = dockercli.RemoveContainer(s.config.DockerHosts[1], s.containerName)
+	out, err = dockercli.RemoveContainer(s.config.DockerHosts[1], s.containerName[0])
 	c.Assert(err, IsNil, Commentf(out))
 
 	// 5. Verify a container can be started to use the same volume on another host
-	out, err = dockercli.AttachVolume(s.config.DockerHosts[0], s.volumeName, s.containerName)
+	out, err = dockercli.AttachVolume(s.config.DockerHosts[0], s.volumeName, s.containerName[0])
 	c.Assert(err, IsNil, Commentf(out))
 
 	// 6. Restart the docker daemon on the other host
@@ -99,7 +119,7 @@ func (s *RestartTestData) TestVolumeDetached(c *C) {
 	c.Assert(err, IsNil, Commentf(out))
 
 	// Clean up the container on host0
-	out, err = dockercli.RemoveContainer(s.config.DockerHosts[0], s.containerName)
+	out, err = dockercli.RemoveContainer(s.config.DockerHosts[0], s.containerName[0])
 	c.Assert(err, IsNil, Commentf(out))
 
 	// 7. Verify detached status for the volume
@@ -107,7 +127,7 @@ func (s *RestartTestData) TestVolumeDetached(c *C) {
 	c.Assert(status, Equals, true, Commentf("Volume %s is still attached", s.volumeName))
 
 	// 8. Verify a container can be started to use the same volume on the original host
-	out, err = dockercli.ExecContainer(s.config.DockerHosts[1], s.volumeName, s.containerName)
+	out, err = dockercli.ExecContainer(s.config.DockerHosts[1], s.volumeName, s.containerName[0])
 	c.Assert(err, IsNil, Commentf(out))
 
 	misc.LogTestEnd(c.TestName())
@@ -125,7 +145,7 @@ func (s *RestartTestData) TestPluginKill(c *C) {
 	misc.LogTestStart(c.TestName())
 
 	// 1. Attach a volume with restart=always flag
-	out, err := dockercli.AttachVolumeWithRestart(s.config.DockerHosts[1], s.volumeName, s.containerName)
+	out, err := dockercli.AttachVolumeWithRestart(s.config.DockerHosts[1], s.volumeName, s.containerName[0])
 	c.Assert(err, IsNil, Commentf(out))
 
 	// 2. Verify volume attached status
@@ -137,15 +157,15 @@ func (s *RestartTestData) TestPluginKill(c *C) {
 	c.Assert(err, IsNil, Commentf(out))
 
 	// 4. Stop the container
-	out, err = dockercli.StopContainer(s.config.DockerHosts[1], s.containerName)
+	out, err = dockercli.StopContainer(s.config.DockerHosts[1], s.containerName[0])
 	c.Assert(err, IsNil, Commentf(out))
 
 	// 5. Start the same container
-	out, err = dockercli.StartContainer(s.config.DockerHosts[1], s.containerName)
+	out, err = dockercli.StartContainer(s.config.DockerHosts[1], s.containerName[0])
 	c.Assert(err, IsNil, Commentf(out))
 
 	// 6. Stop the container
-	out, err = dockercli.StopContainer(s.config.DockerHosts[1], s.containerName)
+	out, err = dockercli.StopContainer(s.config.DockerHosts[1], s.containerName[0])
 	c.Assert(err, IsNil, Commentf(out))
 
 	// 7. Verify volume detached status
@@ -153,7 +173,163 @@ func (s *RestartTestData) TestPluginKill(c *C) {
 	c.Assert(status, Equals, true, Commentf("Volume %s is still attached", s.volumeName))
 
 	// Cleanup container
-	out, err = dockercli.RemoveContainer(s.config.DockerHosts[1], s.containerName)
+	out, err = dockercli.RemoveContainer(s.config.DockerHosts[1], s.containerName[0])
+	c.Assert(err, IsNil, Commentf(out))
+
+	misc.LogTestEnd(c.TestName())
+}
+
+// TestRecoverMountsAfterRestart  - Verify that volume mounts are recovered after
+// docker is restarted and volume is usable by any VM.
+// 1. Start 2 containers using same volume with restart=always
+// 2. Restart docker and wait for containers to start and reference counts to be initialized.
+// 3. Verify volume is attached to the VM
+// 4. Start 1 container using same volume
+// 5. Stop all 3 containers and confirm disk is detached
+// 6. Run container on other host to verify the volume is detached after stopping the containers
+func (s *RestartTestData) TestRecoverMountsAfterRestart(c *C) {
+	misc.LogTestStart(c.TestName())
+
+	c.Skip("Skipping verification of attached status after docker restart")
+	// 1. Run containers with restart-always flag
+	out, err := dockercli.AttachVolumeWithRestart(s.config.DockerHosts[1], s.volumeName, s.containerName[0])
+	c.Assert(err, IsNil, Commentf(out))
+
+	// Run second container 
+	out, err = dockercli.AttachVolumeWithRestart(s.config.DockerHosts[1], s.volumeName, s.containerName[1])
+	c.Assert(err, IsNil, Commentf(out))
+
+	// 2. Restart docker
+	out, err = dockercli.RestartDocker(s.config.DockerHosts[1])
+	c.Assert(err, IsNil, Commentf(out))
+
+	misc.SleepForSec(60)
+
+	// 3. Verify the volume is attached to the VM
+	status, err := dockercli.GetVolumeStatus(s.config.DockerHosts[1], s.volumeName)
+	c.Assert(err, IsNil, Commentf("Failed to fetch status for volume %s", s.volumeName))
+	c.Assert(s.config.DockerHostNames[0], Equals, status["attached_vm"], Commentf("Volume is not attached to VM %s, attached to %s",
+		s.config.DockerHostNames[0], status["attached_vm"]))
+
+	// 4. Run third container
+	out, err = dockercli.AttachVolumeWithRestart(s.config.DockerHosts[1], s.volumeName, s.containerName[2])
+	c.Assert(err, IsNil, Commentf(out))
+
+	// 5. Stop the three containers
+	out, err = dockercli.RemoveContainer(s.config.DockerHosts[1], s.containerName[0])
+	c.Assert(err, IsNil, Commentf(out))
+
+	out, err = dockercli.RemoveContainer(s.config.DockerHosts[1], s.containerName[1])
+	c.Assert(err, IsNil, Commentf(out))
+
+	out, err = dockercli.RemoveContainer(s.config.DockerHosts[1], s.containerName[2])
+	c.Assert(err, IsNil, Commentf(out))
+
+	// 6. Run container on other host to verify the volume is detached after stopping the containers
+	out, err = dockercli.ExecContainer(s.config.DockerHosts[0], s.volumeName, s.containerName[0])
+	c.Assert(err, IsNil, Commentf(out))
+
+	misc.LogTestEnd(c.TestName())
+}
+
+// TestLongVolumeName - Verifies volumes can be created on the default
+// datastore by long and short names.
+// 1. Start container with restart flag and use short name for volume (vol1)
+// 2. Restart docker
+// 3. Start another instance of container with short name for volume (vol1)
+// 4. Get the volume properties and figure the datastore of the volume
+// 4. Start another instance of container with long name for volume (vol1@<datastore-name>)
+// 5. Stop all 3 containers and confirm disk is detached (verify by exec'ing a container on another host).
+func (s *RestartTestData) TestLongVolumeName(c *C) {
+	misc.LogTestStart(c.TestName())
+
+	// 1. Run container with restart-always flag
+	out, err := dockercli.AttachVolumeWithRestart(s.config.DockerHosts[1], s.volumeName, s.containerName[0])
+	c.Assert(err, IsNil, Commentf(out))
+
+	// 2. Restart docker
+	out, err = dockercli.RestartDocker(s.config.DockerHosts[1])
+	c.Assert(err, IsNil, Commentf(out))
+
+	// 3. Run second container 
+	out, err = dockercli.AttachVolumeWithRestart(s.config.DockerHosts[1], s.volumeName, s.containerName[1])
+	c.Assert(err, IsNil, Commentf(out))
+
+	// 3. Get volume status
+	status, err := dockercli.GetVolumeStatus(s.config.DockerHosts[1], s.volumeName)
+	c.Assert(err, IsNil, Commentf("Failed to fetch status for volume %s", s.volumeName))
+
+	// 4. Run third container with long volume name
+	out, err = dockercli.AttachVolumeWithRestart(s.config.DockerHosts[1], s.volumeName + "@" + status["datastore"], s.containerName[2])
+	c.Assert(err, IsNil, Commentf(out))
+
+	// 5. Stop the three containers
+	out, err = dockercli.RemoveContainer(s.config.DockerHosts[1], s.containerName[0])
+	c.Assert(err, IsNil, Commentf(out))
+
+	out, err = dockercli.RemoveContainer(s.config.DockerHosts[1], s.containerName[1])
+	c.Assert(err, IsNil, Commentf(out))
+
+	out, err = dockercli.RemoveContainer(s.config.DockerHosts[1], s.containerName[2])
+	c.Assert(err, IsNil, Commentf(out))
+
+	// 6. Run container on other host to verify the volume is detached after stopping the containers
+	out, err = dockercli.ExecContainer(s.config.DockerHosts[0], s.volumeName, s.containerName[0])
+	c.Assert(err, IsNil, Commentf(out))
+	misc.LogTestEnd(c.TestName())
+}
+
+// TestDuplicateVolumeName - Verifies volumes with same same can be created on the default
+// datastore and non-default datastores
+// 1. Start container with restart flag and short name "vol1" (assuming vol1@ds1)
+// 2. Start container with restart flag with same volume name on another ds: vol1@ds2
+// 3. Restart docker
+// 4. Start container with long name: vol1@ds1
+// 5. Stop all 3 containers and make sure both vol1@ds1 ad vol1@ds2 are detached.
+func (s *RestartTestData) TestDuplicateVolumeName(c *C) {
+	misc.LogTestStart(c.TestName())
+
+	// 1. Run container with restart-always flag
+	out, err := dockercli.AttachVolumeWithRestart(s.config.DockerHosts[1], s.volumeName, s.containerName[0])
+	c.Assert(err, IsNil, Commentf(out))
+
+	// Get volume status
+	status, err := dockercli.GetVolumeStatus(s.config.DockerHosts[1], s.volumeName)
+	c.Assert(err, IsNil, Commentf("Failed to fetch status for volume %s", s.volumeName))
+	
+	ds2 := s.config.Datastores[0]
+	if strings.Compare(status["datastore"], s.config.Datastores[0]) == 0 {
+		ds2 = s.config.Datastores[1]
+	}
+
+	// 2. Run second container with same volume name on the other datastore 
+	out, err = dockercli.AttachVolumeWithRestart(s.config.DockerHosts[1], s.volumeName + "@" + ds2, s.containerName[1])
+	c.Assert(err, IsNil, Commentf(out))
+
+	// 3. Restart docker
+	out, err = dockercli.RestartDocker(s.config.DockerHosts[1])
+	c.Assert(err, IsNil, Commentf(out))
+
+	// 4. Run third container with same volume as the first container
+	out, err = dockercli.AttachVolumeWithRestart(s.config.DockerHosts[1], s.volumeName + "@" + status["datastore"], s.containerName[2])
+	c.Assert(err, IsNil, Commentf(out))
+
+	// 5. Stop the three containers
+	out, err = dockercli.RemoveContainer(s.config.DockerHosts[1], s.containerName[0])
+	c.Assert(err, IsNil, Commentf(out))
+
+	out, err = dockercli.RemoveContainer(s.config.DockerHosts[1], s.containerName[1])
+	c.Assert(err, IsNil, Commentf(out))
+
+	out, err = dockercli.RemoveContainer(s.config.DockerHosts[1], s.containerName[2])
+	c.Assert(err, IsNil, Commentf(out))
+
+	// 6. Run container on other host to verify the volume is detached after stopping the containers
+	out, err = dockercli.ExecContainer(s.config.DockerHosts[0], s.volumeName, s.containerName[0])
+	c.Assert(err, IsNil, Commentf(out))
+
+	// 7. Cleanup volume on the second datastore
+	out, err = dockercli.DeleteVolume(s.config.DockerHosts[1], s.volumeName + "@" + ds2)
 	c.Assert(err, IsNil, Commentf(out))
 
 	misc.LogTestEnd(c.TestName())
